requirements.txt
flask == 2.2.3
google-api-python-client >= 2.82.0
requests~=2.28.2
feedgenerator >= 2.1.0
python-dotenv == 1.0.0